FPS-20 - Выполнен переход на PlatformIO.
Проверка перехода завершена только сборкой, требуется отладка на прототипе.
