+++ platformio.ini
@@ -1,0 +1,20 @@
+; PlatformIO Project Configuration File
+;
+;   Build options: build flags, source filter
+;   Upload options: custom upload port, speed and extra flags
+;   Library options: dependencies, extra library storages
+;   Advanced options: extra scripting
+;
+; Please visit documentation for the other options and examples
+; https://docs.platformio.org/page/projectconf.html
+
+[env:esp32dev]
+platform = espressif32
+board = esp32dev
+framework = arduino
+lib_deps = 
+	adafruit/Adafruit SSD1306@^2.5.13
+	khuuxuanngoc/DHT kxn@^3.4.4
+	adafruit/Adafruit Unified Sensor@^1.1.15
+build_flags =
+	-std=c++17
