--- src/Screen/BaseFunctionalScreen.cpp
@@ -1,6 +1,0 @@
-#include "BaseFunctionalScreen.h"
-
-void TemperatureSensorScreen::printScreen(std::shared_ptr<Adafruit_SSD1306> display)
-{
-    //TODO: Реализовать вывод на экран температуры и прочих параметров
-}
