+++ src/Screen/BaseFunctionalScreens/BaseFunctionalScreen.cpp
@@ -1,0 +1,2 @@
+#include "BaseFunctionalScreen.h"
+
