+++ src/Screen/BaseFunctionalScreens/BaseFunctionalScreen.h
@@ -19,13 +19,3 @@ public:
 };
 
 
-class TemperatureSensorScreen : public BaseFunctionalScreen
-{
-
-public:
-    TemperatureSensorScreen() = default;
-    ~TemperatureSensorScreen() = default;
-
-    void printScreen(std::shared_ptr<Adafruit_SSD1306> display) override;
-};
-
